User flow update
Added some html and tools so we can implement a new user flow.
The modal is not in use right now.
We will need to refactor this modal to it's own main page which will seem to work like a modal but only act like a welcome message and redirect the users to either view the map or, login to contribute.